Use int instead of byte for scenarios

+++ src/edu/uclouvain/core/nodus/compute/assign/AssignmentParameters.java
@@ -92,7 +92,7 @@ public class AssignmentParameters {
   private boolean savePaths;
 
   /** Scenario to assign. */
-  private byte scenario = -1;
+  private int scenario = -1;
 
   /** The number of parallel threads during assignment. */
   private int threads = 1;
@@ -204,9 +204,9 @@ public class AssignmentParameters {
   /**
    * Returns the num of the assigned scenario.
    *
-   * @return byte.
+   * @return int.
    */
-  public byte getScenario() {
+  public int getScenario() {
     return scenario;
   }
 
@@ -499,10 +499,10 @@ public class AssignmentParameters {
   /**
    * Sets the num of the scenario corresponding to this assignment.
    *
-   * @param scenario byte
+   * @param scenario int
    */
-  public void setScenario(byte scenario) {
-    this.scenario = (byte) scenario;
+  public void setScenario(int scenario) {
+    this.scenario = scenario;
   }
 
   /**
