+++ src/edu/uclouvain/core/nodus/compute/assign/DynamicTimeDependentAssignment.java
@@ -108,7 +108,7 @@ public class DynamicTimeDependentAssignment extends Assignment {
       return false;
     }
 
-    byte scenario = assignmentParameters.getScenario();
+    int scenario = assignmentParameters.getScenario();
 
     // Read the exclusions
     NodeRulesReader er = new NodeRulesReader(virtualNet, scenario);
