+++ src/edu/uclouvain/core/nodus/compute/assign/ExactMFAssignment.java
@@ -83,7 +83,7 @@ public class ExactMFAssignment extends Assignment {
       return false;
     }
 
-    byte scenario = assignmentParameters.getScenario();
+    int scenario = assignmentParameters.getScenario();
 
     // Read the exclusions
     NodeRulesReader er = new NodeRulesReader(virtualNet, scenario);
