+++ src/edu/uclouvain/core/nodus/compute/assign/FrankWolfeAssignment.java
@@ -83,7 +83,7 @@ public class FrankWolfeAssignment extends Assignment {
       return false;
     }
 
-    byte scenario = assignmentParameters.getScenario();
+    int scenario = assignmentParameters.getScenario();
 
     // Read the exclusions
     NodeRulesReader er = new NodeRulesReader(virtualNet, scenario);
